add -p option to control which data page gets fetched
Note that this feature does not work yet. It seems that the data
page that comes back has references to a session id that does not
match the one that we pass in. Odd. My current guess is that we
must re-use the TCP connection that we use to create the session.